requirements.txt
ipykernel
shiny
pytest
pytest-playwright
shiny_adaptive_filter
